refactor(my-diary): rename components to fragments, add fetchInitialData file
affects: @my-diary/app, @fragments/form-add-post, @fragments/main-page-header,
@fragments/post-controls, @fragments/post-list, @practice/my-diary, @ui/list

ISSUES CLOSED: #77
