+++ practice/my-diary/app/src/App.tsx
@@ -2,10 +2,10 @@ import React, { useState }     from 'react'
 import { ThemeProvider }       from '@emotion/react'
 
 import * as theme              from '@ui/theme'
-import { FormAddPost }         from '@components/form-add-post'
-import { MainPageHeader }      from '@components/main-page-header'
-import { PostControls }        from '@components/post-controls'
-import { PostList }            from '@components/post-list'
+import { FormAddPost }         from '@fragments/form-add-post'
+import { MainPageHeader }      from '@fragments/main-page-header'
+import { PostControls }        from '@fragments/post-controls'
+import { PostList }            from '@fragments/post-list'
 import { FilterProvider }      from '@store/filter'
 import { NotesProvider }       from '@store/notes'
 import { SearchValueProvider } from '@store/search-value'
@@ -13,27 +13,7 @@ import { ShowDrawerProvider }  from '@store/show-drawer'
 import { Box, Column, Layout } from '@ui/layout'
 
 export const App = () => {
-  const [notes, setNotes] = useState([
-    {
-      note: 'Flight to Moscow',
-      important: false,
-      liked: false,
-      done: false,
-    },
-    {
-      note: 'Friends meeting',
-      important: false,
-      liked: false,
-      done: false,
-    },
-    {
-      note: 'Buy a new frying pan in Ikea',
-      important: false,
-      liked: false,
-      done: false,
-    },
-  ])
-
+  const [notes, setNotes] = useState([])
   const [searchValue, setSearchValue] = useState('')
   const [filter, setFilter] = useState('all')
   const [showDrawer, setShowDrawer] = useState<boolean>(null)
