--- practice/my-diary/components/form-add-post/src/actions/index.ts
@@ -1,1 +1,0 @@
-export * from './addItem'
