--- practice/my-diary/components/post-list/src/actions/addId.ts
@@ -1,13 +1,0 @@
-export const addId = notes => {
-  let newNotes = []
-  notes.forEach(item => {
-    newNotes = newNotes.concat({
-      note: item.note,
-      important: item.important,
-      liked: item.liked,
-      done: item.done,
-      id: Math.random() * Date.now(),
-    })
-  })
-  return newNotes
-}
