+++ practice/my-diary/fragments/form-add-post/src/FormAddPost.tsx
@@ -1,20 +1,25 @@
-import React, { useState }     from 'react'
-import { useIntl }             from 'react-intl'
+import React, { useEffect, useState } from 'react'
+import { useIntl }                    from 'react-intl'
 
-import { Button }              from '@ui/button'
-import { Input }               from '@ui/input'
-import { Column, Layout, Row } from '@ui/layout'
-import { Space }               from '@ui/text'
-import { useNotes }            from '@store/notes'
+import { Button }                     from '@ui/button'
+import { Input }                      from '@ui/input'
+import { Column, Layout, Row }        from '@ui/layout'
+import { Space }                      from '@ui/text'
+import { useNotes }                   from '@store/notes'
 
-import messages                from './messages'
-import { addItem }             from './actions'
+import messages                       from './messages'
+import { fetchInitialData }           from './actions'
 
 export const FormAddPost = () => {
   const [newNote, setNewNote] = useState('')
   const [notes, setNotes] = useNotes()
   const [inputStatus, setInputStatus] = useState('invisible')
   const intl = useIntl()
+  const initialNotes = ['Flight to Moscow', 'Friends meeting', 'Buy a new frying pan in Ikea']
+
+  useEffect(() => {
+    setNotes(fetchInitialData(notes, initialNotes))
+  }, [])
   return (
     <Column>
       <Column backgroundColor='white'>
@@ -46,7 +51,7 @@ export const FormAddPost = () => {
         onClick={() => {
           setInputStatus('visible')
           if (inputStatus === 'visible' && newNote !== '') {
-            setNotes(addItem(notes, newNote))
+            setNotes(fetchInitialData(notes, [newNote]))
             setNewNote('')
             setInputStatus('invisible')
           } else if (inputStatus === 'visible') {
