+++ practice/my-diary/fragments/main-page-header/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@components/main-page-header",
+  "name": "@fragments/main-page-header",
   "version": "0.0.0",
   "license": "MIT",
   "main": "src/index.ts",
