+++ practice/my-diary/fragments/main-page-header/src/index.ts
