+++ practice/my-diary/fragments/post-controls/package.json
@@ -1,19 +1,22 @@
 {
-  "name": "@components/form-add-post",
+  "name": "@fragments/post-controls",
   "version": "0.0.0",
   "license": "MIT",
   "main": "src/index.ts",
   "dependencies": {
     "@emotion/react": "^11.1.5",
     "@emotion/styled": "^11.1.5",
-    "react-intl": "^5.15.7",
+    "@store/filter": "0.0.0",
+    "@store/notes": "0.0.0",
+    "@store/search-value": "0.0.0",
     "@ui/button": "0.0.0",
-    "@ui/icons": "0.0.0",
+    "@ui/drawer": "0.0.0",
     "@ui/input": "0.0.0",
     "@ui/layout": "^0.0.0",
+    "@store/show-drawer": "0.0.0",
     "@ui/text": "^0.0.0",
-    "@ui/theme": "0.0.0",
+    "@ui/theme": "^0.0.0",
     "react": "^17.0.1",
-    "@store/notes": "0.0.0"
+    "react-intl": "^5.15.7"
   }
 }
