+++ practice/my-diary/fragments/post-list/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@components/post-list",
+  "name": "@fragments/post-list",
   "version": "0.0.0",
   "license": "MIT",
   "main": "src/index.ts",
