+++ practice/my-diary/fragments/post-list/src/actions/deleteItem.ts
