+++ practice/my-diary/fragments/post-list/src/actions/index.ts
@@ -1,4 +1,3 @@
 export * from './filterNotes'
 export * from './deleteItem'
 export * from './toggleStatus'
-export * from './addId'
