+++ practice/my-diary/fragments/post-list/src/actions/toggleStatus.ts
